properly quote bound Vue props (using v-bind shorthand)
This change allows for dynamically bound Vue props that are missing their quotes to be properly quoted.  Performing the change in `src/printer.ts` fixed the bug, but caused other tests to fail.  Adding the additional back-tic quote in `isQuoted` allows all unit tests to pass.  I looked at all of the calling sites of `isQuoted` and it looks okay to me, but it should probably be double checked.